Adicionado objetos, vehicle, entrance, protocol, policeData, entranceServiceBean, protocolServiceBean
