+++ src/main/java/br/com/patiolegal/domain/PoliceData.java
@@ -1,0 +1,96 @@
+package br.com.patiolegal.domain;
+
+public class PoliceData {
+
+	private Boolean insured;
+	private Boolean financed;
+	private Boolean stolen;
+	private Boolean drugTrafficking;
+	private Boolean moneyLaundry;
+	private Boolean perquisite;
+	private Boolean papillaryExpertise;
+	private Boolean ownerIntimate;
+	private Boolean authorizedAlienation;
+	private Boolean debits;
+
+	public Boolean getInsured() {
+		return insured;
+	}
+
+	public Boolean getFinanced() {
+		return financed;
+	}
+
+	public Boolean getStolen() {
+		return stolen;
+	}
+
+	public Boolean getDrugTrafficking() {
+		return drugTrafficking;
+	}
+
+	public Boolean getMoneyLaundry() {
+		return moneyLaundry;
+	}
+
+	public Boolean getPerquisite() {
+		return perquisite;
+	}
+
+	public Boolean getPapillaryExpertise() {
+		return papillaryExpertise;
+	}
+
+	public Boolean getOwnerIntimate() {
+		return ownerIntimate;
+	}
+
+	public Boolean getAuthorizedAlienation() {
+		return authorizedAlienation;
+	}
+
+	public Boolean getDebits() {
+		return debits;
+	}
+
+	public void setInsured(Boolean insured) {
+		this.insured = insured;
+	}
+
+	public void setFinanced(Boolean financed) {
+		this.financed = financed;
+	}
+
+	public void setStolen(Boolean stolen) {
+		this.stolen = stolen;
+	}
+
+	public void setDrugTrafficking(Boolean drugTrafficking) {
+		this.drugTrafficking = drugTrafficking;
+	}
+
+	public void setMoneyLaundry(Boolean moneyLaundry) {
+		this.moneyLaundry = moneyLaundry;
+	}
+
+	public void setPerquisite(Boolean perquisite) {
+		this.perquisite = perquisite;
+	}
+
+	public void setPapillaryExpertise(Boolean papillaryExpertise) {
+		this.papillaryExpertise = papillaryExpertise;
+	}
+
+	public void setOwnerIntimate(Boolean ownerIntimate) {
+		this.ownerIntimate = ownerIntimate;
+	}
+
+	public void setAuthorizedAlienation(Boolean authorizedAlienation) {
+		this.authorizedAlienation = authorizedAlienation;
+	}
+
+	public void setDebits(Boolean debits) {
+		this.debits = debits;
+	}
+
+}
