+++ src/main/java/br/com/patiolegal/domain/ShedData.java
@@ -1,0 +1,39 @@
+package br.com.patiolegal.domain;
+
+public class ShedData {
+	
+	private String shed;
+	private String row;
+	private String column;
+	private String floor;
+	
+	public String getShed() {
+		return shed;
+	}
+	public String getRow() {
+		return row;
+	}
+	public String getColumn() {
+		return column;
+	}
+	public String getFloor() {
+		return floor;
+	}
+	
+	public void setShed(String shed) {
+		this.shed = shed;
+	}
+
+	public void setRow(String row) {
+		this.row = row;
+	}
+
+	public void setColumn(String column) {
+		this.column = column;
+	}
+
+	public void setFloor(String floor) {
+		this.floor = floor;
+	}	
+
+}
