+++ src/main/java/br/com/patiolegal/domain/Vehicle.java
@@ -2,32 +2,149 @@ package br.com.patiolegal.domain;
 
 public class Vehicle {
 
-    private String theyRenamed;
-    private String sportingPlate;
-    private String originalPlate;
+	private String theyRenamed;
+	private String sportingPlate;
+	private String originalPlate;
+	private String ownerName;
+	private String ownerTaxIdentifier;
+	private String brand;
+	private String model;
+	private String category;
+	private String color;
+	private String fuel;
+	private Integer yearFactory;
+	private Integer yearModel;
+	private String chassisState;
+	private String chassis;
+	private String motorState;
+	private String motor;
 
-    public String getSportingPlate() {
-        return sportingPlate;
-    }
+	public String getTheyRenamed() {
+		return theyRenamed;
+	}
 
-    public void setSportingPlate(String sportingPlate) {
-        this.sportingPlate = sportingPlate;
-    }
+	public void setTheyRenamed(String theyRenamed) {
+		this.theyRenamed = theyRenamed;
+	}
 
-    public String getOriginalPlate() {
-        return originalPlate;
-    }
+	public String getSportingPlate() {
+		return sportingPlate;
+	}
 
-    public void setOriginalPlate(String originalPlate) {
-        this.originalPlate = originalPlate;
-    }
+	public void setSportingPlate(String sportingPlate) {
+		this.sportingPlate = sportingPlate;
+	}
 
-    public String getTheyRenamed() {
-        return theyRenamed;
-    }
+	public String getOriginalPlate() {
+		return originalPlate;
+	}
 
-    public void setTheyRenamed(String theyRenamed) {
-        this.theyRenamed = theyRenamed;
-    }
+	public void setOriginalPlate(String originalPlate) {
+		this.originalPlate = originalPlate;
+	}
+
+	public String getOwnerName() {
+		return ownerName;
+	}
+
+	public void setOwnerName(String ownerName) {
+		this.ownerName = ownerName;
+	}
+
+	public String getOwnerTaxIdentifier() {
+		return ownerTaxIdentifier;
+	}
+
+	public void setOwnerTaxIdentifier(String ownerTaxIdentifier) {
+		this.ownerTaxIdentifier = ownerTaxIdentifier;
+	}
+
+	public String getBrand() {
+		return brand;
+	}
+
+	public void setBrand(String brand) {
+		this.brand = brand;
+	}
+
+	public String getModel() {
+		return model;
+	}
+
+	public void setModel(String model) {
+		this.model = model;
+	}
+
+	public String getCategory() {
+		return category;
+	}
+
+	public void setCategory(String category) {
+		this.category = category;
+	}
+
+	public String getColor() {
+		return color;
+	}
+
+	public void setColor(String color) {
+		this.color = color;
+	}
+
+	public String getFuel() {
+		return fuel;
+	}
+
+	public void setFuel(String fuel) {
+		this.fuel = fuel;
+	}
+
+	public Integer getYearFactory() {
+		return yearFactory;
+	}
+
+	public void setYearFactory(Integer yearFactory) {
+		this.yearFactory = yearFactory;
+	}
+
+	public Integer getYearModel() {
+		return yearModel;
+	}
+
+	public void setYearModel(Integer yearModel) {
+		this.yearModel = yearModel;
+	}
+
+	public String getChassisState() {
+		return chassisState;
+	}
+
+	public void setChassisState(String chassisState) {
+		this.chassisState = chassisState;
+	}
+
+	public String getChassis() {
+		return chassis;
+	}
+
+	public void setChassis(String chassis) {
+		this.chassis = chassis;
+	}
+
+	public String getMotorState() {
+		return motorState;
+	}
+
+	public void setMotorState(String motorState) {
+		this.motorState = motorState;
+	}
+
+	public String getMotor() {
+		return motor;
+	}
+
+	public void setMotor(String motor) {
+		this.motor = motor;
+	}
 
 }
