+++ src/main/java/br/com/patiolegal/service/EntranceServiceBean.java
@@ -30,6 +30,20 @@ public class EntranceServiceBean implements EntranceService {
         Entrance entrance = new Entrance();
         Vehicle vehicle = new Vehicle();
         vehicle.setOriginalPlate(request.getOriginalPlate());
+        vehicle.setSportingPlate(request.getSportingPlate());
+        vehicle.setOwnerName(request.getOwnerName());
+        vehicle.setOwnerTaxIdentifier(request.getOwnerTaxIdentifier());
+        vehicle.setBrand(request.getBrand());
+        vehicle.setModel(request.getModel());
+        vehicle.setCategory(request.getCategory());
+        vehicle.setColor(request.getColor());
+        vehicle.setFuel(request.getFuel());
+        vehicle.setYearFactory(request.getYearFactory());
+        vehicle.setYearModel(request.getYearModel());
+        vehicle.setChassisState(request.getChassisState());
+        vehicle.setChassis(request.getChassis());
+        vehicle.setMotorState(request.getMotorState());
+        vehicle.setMotor(request.getMotor());
         entrance.setVehicle(vehicle);
         protocol.setEntrance(entrance);
         entranceRepository.save(protocol);
