+++ src/main/java/br/com/patiolegal/service/ProtocolServiceBean.java
@@ -42,36 +42,36 @@ public class ProtocolServiceBean implements ProtocolService{
 										.withTaxId(protocolValue.getTaxId())
 										.withName(protocolValue.getName())
 										.withTheyRenamed(protocolValue.getEntrance().getVehicle().getTheyRenamed())
-										.withOwnerName(protocolValue.getOwnerName())
-										.withOwnerTaxIdentifier(protocolValue.getOwnerTaxIdentifier())
-										.withBrand(protocolValue.getBrand())
-										.withModel(protocolValue.getModel())
-										.withCategory(protocolValue.getCategory())
-										.withColor(protocolValue.getColor())
-										.withFuel(protocolValue.getFuel())
-										.withYearFactory(protocolValue.getYearFactory())
-										.withYearModel(protocolValue.getYearModel())
+										.withOwnerName(protocolValue.getEntrance().getVehicle().getOwnerName())
+										.withOwnerTaxIdentifier(protocolValue.getEntrance().getVehicle().getOwnerTaxIdentifier())
+										.withBrand(protocolValue.getEntrance().getVehicle().getBrand())
+										.withModel(protocolValue.getEntrance().getVehicle().getModel())
+										.withCategory(protocolValue.getEntrance().getVehicle().getCategory())
+										.withColor(protocolValue.getEntrance().getVehicle().getColor())
+										.withFuel(protocolValue.getEntrance().getVehicle().getFuel())
+										.withYearFactory(protocolValue.getEntrance().getVehicle().getYearFactory())
+										.withYearModel(protocolValue.getEntrance().getVehicle().getYearModel())
 										.withSportingPlate(protocolValue.getEntrance().getVehicle().getSportingPlate())
 										.withOriginalPlate(protocolValue.getEntrance().getVehicle().getOriginalPlate())
 										.withOriginCapture(protocolValue.getOriginCapture())
-										.withChassisState(protocolValue.getChassisState())
-										.withChassis(protocolValue.getChassis())
-										.withMotorState(protocolValue.getMotorState())
-										.withMotor(protocolValue.getMotor())
-										.withInsured(protocolValue.getInsured())
-										.withFinanced(protocolValue.getFinanced())
-										.withStolen(protocolValue.getStolen())
-										.withDrugTrafficking(protocolValue.getDrugTrafficking())
-										.withMoneyLaundry(protocolValue.getMoneyLaundry())
-										.withPerquisite(protocolValue.getPerquisite())
-										.withPapillaryExpertise(protocolValue.getPapillaryExpertise())
-										.withOwnerIntimate(protocolValue.getOwnerIntimate())
-										.withAuthorizedAlienation(protocolValue.getAuthorizedAlienation())
-										.withDebits(protocolValue.getDebits())
-										.withShed(protocolValue.getShed())
-										.withRow(protocolValue.getRow())
-										.withColumn(protocolValue.getColumn())
-										.withFloor(protocolValue.getFloor())
+										.withChassisState(protocolValue.getEntrance().getVehicle().getChassisState())
+										.withChassis(protocolValue.getEntrance().getVehicle().getChassis())
+										.withMotorState(protocolValue.getEntrance().getVehicle().getMotorState())
+										.withMotor(protocolValue.getEntrance().getVehicle().getMotor())
+										.withInsured(protocolValue.getEntrance().getPoliceData().getInsured())
+										.withFinanced(protocolValue.getEntrance().getPoliceData().getFinanced())
+										.withStolen(protocolValue.getEntrance().getPoliceData().getStolen())
+										.withDrugTrafficking(protocolValue.getEntrance().getPoliceData().getDrugTrafficking())
+										.withMoneyLaundry(protocolValue.getEntrance().getPoliceData().getMoneyLaundry())
+										.withPerquisite(protocolValue.getEntrance().getPoliceData().getPerquisite())
+										.withPapillaryExpertise(protocolValue.getEntrance().getPoliceData().getPapillaryExpertise())
+										.withOwnerIntimate(protocolValue.getEntrance().getPoliceData().getOwnerIntimate())
+										.withAuthorizedAlienation(protocolValue.getEntrance().getPoliceData().getAuthorizedAlienation())
+										.withDebits(protocolValue.getEntrance().getPoliceData().getDebits())
+										.withShed(protocolValue.getEntrance().getShedData().getShed())
+										.withRow(protocolValue.getEntrance().getShedData().getRow())
+										.withColumn(protocolValue.getEntrance().getShedData().getColumn())
+										.withFloor(protocolValue.getEntrance().getShedData().getFloor())
 										.withAccountableOut(protocolValue.getAccountableOut())
 										.withAccountableIn(protocolValue.getAccountableIn())
 										.withBoard(protocolValue.getBoard())
